nH_MWC_Linkage
This directory contains the explicit mathemathical expressions for the dependence of [S0.5] and nH at half saturation on the elementary Kr, Kt and L MWC parameters. The [S0.5] expression was obtained by setting equation 2 in the main text equals 1/2. Explicit solution was obtained using the symbolic toolbox of the (MATLAB R2015b) software package (Symbolic Toolbox Release 2015b, The MathWorks, Inc., Natick, Massachusetts, United States). Substituting the expression for [S0.5] (below) in equation 5 of the main text  then allowed deriving the explicit expression of nH (see further below). please refer to the main text of the manuscript for further details.